02_walk/requirements.txt
httpx
python-dotenv

docling
numpy
cohere

lancedb
pydantic
opentelemetry-api==1.39.1
opentelemetry-exporter-gcp-logging==1.10.0a0
opentelemetry-exporter-otlp-proto-http==1.39.1

google-adk==1.14.1
google-cloud-aiplatform==1.126.1
google-cloud-discoveryengine
google-cloud-storage==2.19.0
